% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{babynames}
\alias{babynames}
\title{Popular names data}
\format{
\subsection{\code{babynames}}{

A data frame with 31915 rows and 3 columns:
\describe{
\item{name}{Baby's name}
\item{gender}{F if female, M if male}
\item{count}{Number of babies with name and gender}
}
}
}
\source{
\url{https://www.ssa.gov/oact/babynames/limits.html}
}
\usage{
babynames
}
\description{
Data on the names of all babies born in the United States in 2022, as
provided by the Social Security Administration. Each observation corresponds
to a specific name and gender, with a count of that name provided. For
confidentiality reasons, the minimum count for any name is 5. All other
names (with fewer than 5 occurrences in the U.S.) are included within the
observation having "OTHER" as the name. There are two "OTHER" observations,
one for female babies and one for male babies. Data are sorted alphabetically
by name.
}
\keyword{datasets}
