% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baseball}
\alias{baseball}
\title{Baseball attendance data}
\format{
\subsection{\code{baseball}}{

A data frame with 30 rows and 9 columns:
\describe{
\item{team}{Team name}
\item{attend_home}{Average home game attendance}
\item{attend_road}{Average road game attendance}
\item{winpct_22}{Team winning percentage in 2022}
\item{winpct_21}{Team winning percentage in 2021}
\item{playoff_21}{1 if team made playoffs in 2021, 0 otherwise}
\item{capacity}{Capacity of home stadium}
\item{popul}{Population of team's metropolitan area (2020)}
\item{payroll}{Total team payroll in 2022 (in millions of dollars)}
}
}
}
\source{
various
}
\usage{
baseball
}
\description{
Data on 2022 attendance for Major League Baseball teams
}
\keyword{datasets}
