% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bitcoin}
\alias{bitcoin}
\title{Bitcoin price and returns data}
\format{
\subsection{\code{bitcoin}}{

A data frame with 364 rows and 268 columns:
\describe{
\item{date}{Date}
\item{high}{Highest price (in dollars)}
\item{low}{Lowest price (in dollars)}
\item{close}{End-of-day price (in dollars)}
\item{return}{Daily return, based on end-of-day prices}
}
}
}
\source{
\url{https://finance.yahoo.com}
}
\usage{
bitcoin
}
\description{
Data on daily prices and returns for Bitcoin during 2020 and 2021
}
\keyword{datasets}
