% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brands}
\alias{brands}
\title{Brand data}
\format{
\subsection{\code{brands}}{

A data frame with 14,560 rows and 3 columns:
\describe{
\item{purchase}{1 if customer makes a purchase, 0 otherwise}
\item{brand}{Brand purchased (1 through 5), 0 if no purchase}
\item{last_brand}{Brand purchased (1 through 5) during last visit}
}
}
}
\source{
\url{https://medium.com/\%40miradzji/purchase-probability-analysis-in-certain-market-segments-with-python-b346654ea5ec}
}
\usage{
brands
}
\description{
Data on the purchase behavior of customers at a specific market. The dataset
consists of customers who purchased one of five candy-bar brands in their previous
visit to the market and records whether or not they make a purchase during
this visit and, if so, which brand they purchase. The dataset is adapted from
the full dataset that is referenced in the source citation.
}
\keyword{datasets}
