% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cigdata}
\alias{cigdata}
\title{State-level cigarette price and tax data}
\format{
\subsection{\code{cigdata}}{

A data frame with 51 rows and 9 columns:
\describe{
\item{state}{State abbreviation}
\item{statename}{State name}
\item{cigprice}{Average price per pack (in dollars)}
\item{cigsales}{Annual sales, packs per capita}
\item{cig_tax_revenue}{Total annual tax revenue (in dollars)}
\item{cigtax}{State tax per pack (in dollars)}
\item{producer}{1 if tobacco production > 20m pounds, 0 otherwise}
}
}
}
\source{
\url{https://healthdata.gov/dataset/The-Tax-Burden-on-Tobacco-1970-2019/etts-u9ii}
}
\usage{
cigdata
}
\description{
Data on cigarette prices and taxes in 2019 for the 50 U.S. states plus the
District of Columbia
}
\keyword{datasets}
