% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cps}
\alias{cps}
\title{Current Population Survey (CPS) data}
\format{
\subsection{\code{cps}}{

A data frame with 4,013 rows and 17 columns:
\describe{
\item{statefips}{Two-character state code, including DC}
\item{gender}{Gender (Male, Female)}
\item{metro}{Metropolitan-area (Metro, Non-Metro)}
\item{race}{Race category (Black, White, Other)}
\item{hispanic}{Hispanic (Hispanic, Non-hispanic)}
\item{marstatus}{Marital status (Married, Divorced, Widowed, Never married)}
\item{lfstatus}{Labor-force status (Employed, Unemployed, Not in LF)}
\item{ottipcomm}{Earnings include overtime, tips, and/or commissions (Yes, No)}
\item{hourly}{Hourly-worker status (Hourly, Non-hourly)}
\item{unionstatus}{Union status (Union, Non-union)}
\item{age}{Age (in years)}
\item{hrslastwk}{Hours worked last week}
\item{unempwks}{Number of weeks unemployed}
\item{wagehr}{Hourly wage (in dollars); only for hourly employees}
\item{earnwk}{Earnings last week (in dollars)}
\item{ownchild}{Number of children in household}
\item{educ}{Highest education level attained (in years)}
}
}
}
\source{
\url{https://www.census.gov/programs-surveys/cps/data/datasets.html}
}
\usage{
cps
}
\description{
A subsample of the 2019 Current Population Survey (CPS) consisting of
data on individuals aged 30 to 59 (inclusive)
}
\keyword{datasets}
