% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dictator}
\alias{dictator}
\title{Dictator-game data}
\format{
\subsection{\code{dictator}}{

A data frame with 137 rows and 5 columns:
\describe{
\item{earnings}{Earnings of the dictator (between 0 and 10)}
\item{giving}{1 if giving game, 0 otherwise}
\item{taking}{1 if taking game, 0 otherwise}
\item{female}{1 if dictator is female, 0 otherwise}
\item{female_opp}{1 if recipient is female, 0 otherwise}
}
}
}
\source{
\url{https://journaldata.zbw.eu/dataset/giving-and-taking-in-dictator-games-replication}
}
\usage{
dictator
}
\description{
Data on the results from "dictator games" played in an experimental study, based
on the paper "Giving and taking in dictator games -- differences by gender?
A replication study of Chowdhury et al.", Journal of Comments and Replications
in Economics, 2023. Each observation corresponds to one play of the game.
Earnings are for the dictator. Two game variants are the "giving game"
(dictator starts with endowment) and "taking game" (recipient starts with
endowment).
}
\keyword{datasets}
