% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exams}
\alias{exams}
\title{Exam data}
\format{
\subsection{\code{exams}}{

A data frame with 77 rows and 2 columns:
\describe{
\item{exam1}{Score (out of 100) on the first exam}
\item{exam2}{Score (out of 100) on the second exam}
}
}
}
\usage{
exams
}
\description{
Data on two exam scores for 77 university students
}
\keyword{datasets}
