% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{houseprices}
\alias{houseprices}
\title{Housing price data}
\format{
\subsection{\code{houseprices}}{

A data frame with 973 rows and 16 columns:
\describe{
\item{lotarea}{Area of lot (in square feet)}
\item{overallqual}{Overall home quality (scale 1-10, 10 best)}
\item{yearbuilt}{Year house was built}
\item{yearremodadd}{Year house was remodeled (equal to yearbuilt if never)}
\item{bsmtfinsf}{Area of finished basement (in square feet, 0 if no finished basement)}
\item{grlivarea}{Total non-basement living area (in square feet)}
\item{fullbath}{Number of full bathrooms}
\item{halfbath}{Number of half bathrooms}
\item{bedroomabvgr}{Number of non-basement bedrooms}
\item{totrmsabvgrd}{Number of non-basement rooms (not including bathrooms)}
\item{fireplaces}{Number of fireplaces}
\item{garagecars}{Size of garage (0 if no garage)}
\item{mosold}{Month house sold (1=Jan,...,12=Dec)}
\item{yrsold}{Year house sold}
\item{saleprice}{Sales price of house (in dollars)}
\item{centralair}{1 if house has central air, 0 otherwise}
}
}
}
\source{
\url{https://www.kaggle.com/competitions/house-prices-advanced-regression-techniques/data}
}
\usage{
houseprices
}
\description{
Data on house sales in Ames, Iowa between 2006 and 2010. The dataset is limited
to one-family homes with public utilities and excludes new home sales.
}
\keyword{datasets}
