% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inflation}
\alias{inflation}
\title{Inflation data}
\format{
\subsection{\code{inflation}}{

A data frame with 450 rows and 3 columns:
\describe{
\item{country}{Country abbreviation}
\item{year}{Year}
\item{inflation}{Annual inflation rate (change in CPI)}
}
}
}
\source{
\url{https://data.oecd.org/price/inflation-cpi.htm}
}
\usage{
inflation
}
\description{
Data on inflation rates for 45 countries for a ten-year period (2010-2019).
}
\keyword{datasets}
