% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{married}
\alias{married}
\title{Married-couple data}
\format{
\subsection{\code{married}}{

A data frame with 4,126 rows and 11 columns:
\describe{
\item{age_w}{Age of wife (in years)}
\item{age_h}{Age of husband (in years)}
\item{educ_w}{Education of wife (in years)}
\item{educ_h}{Education of husband (in years)}
\item{bmi_w}{Body mass index of wife (bottom-coded at 18, top-coded at 40)}
\item{bmi_h}{Body mass index of husband (bottom-coded at 18, top-coded at 40)}
\item{smoke_w}{1 if wife smokes, 0 otherwise}
\item{smoke_h}{1 if husband smokes, 0 otherwise}
\item{employed_w}{1 if wife employed, 0 otherwise}
\item{employed_h}{1 if husband employed, 0 otherwise}
\item{famincome}{Annual family income (in dollars, top-coded at $150,000)}
}
}
}
\source{
\url{https://www.icpsr.umich.edu/web/HMCA/studies/4216}
}
\usage{
married
}
\description{
Data on married couples in the United States from the 2003 Community Tracking
Study (CTS) Household Survey.
}
\keyword{datasets}
