% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metricsgrades}
\alias{metricsgrades}
\title{Econometrics course data}
\format{
\subsection{\code{metricsgrades}}{

A data frame with 68 rows and 4 columns:
\describe{
\item{gre_quant}{Score on GRE quantitative test (out of 170)}
\item{gre_verbal}{Score on GRE verbal test (out of 170)}
\item{domestic}{1 if domestic student, 0 if international student}
\item{total}{Overall composite course grade (out of 100 points)}
}
}
}
\usage{
metricsgrades
}
\description{
Data on performance in a graduate econometrics course, with GRE test information
and domestic/international status available.
}
\keyword{datasets}
