% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mutualfunds}
\alias{mutualfunds}
\title{Mutual-fund performance data}
\format{
\subsection{\code{mutualfunds}}{

A data frame with 208 rows and 11 columns:
\describe{
\item{name}{Name of mutual fund}
\item{fund_age}{Age of fund (in years)}
\item{expense_ratio}{Expense ratio (net)}
\item{aum}{Assets under management (in millions of dollars)}
\item{min_investment}{Minimum investment level (in dollars)}
\item{load}{Y if fund has a load (sales charge or fee), N if not}
\item{manager_tenure}{Tenure of current fund manager (in years)}
\item{return_1yr}{One-year annualized return}
\item{return_3yr}{Three-year annualized return}
\item{return_5yr}{Five-year annualized return}
\item{return_10yr}{Ten-year annualized return}
}
}
}
\source{
\url{https://www.fidelity.com}
}
\usage{
mutualfunds
}
\description{
Data on mutual funds categorized as "Large Blend Equity" funds by Morningstar,
limited to funds in existence for more than 10 years. Data captured 2/28/2023.
}
\keyword{datasets}
