% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{premier}
\alias{premier}
\title{Premier League soccer data}
\format{
\subsection{\code{premier}}{

A data frame with 380 rows and 5 columns:
\describe{
\item{round}{Round (values 1 to 38)}
\item{hometeam}{Home team}
\item{awayteam}{Away team}
\item{homegoals}{Number of goals by the home team}
\item{awaygoals}{Number of goals by the away team}
}
}
}
\source{
\url{https://en.wikipedia.org/wiki/2020\%E2\%80\%9321_Premier_League}
}
\usage{
premier
}
\description{
Data on all game results for the 2020 Premier League soccer season. The
Premier League consists of 20 teams. Each team plays every other team twice
(home and away) during the season, so there are a total of 38 rounds in
the season and 380 total games.
}
\keyword{datasets}
