% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{resume}
\alias{resume}
\title{Resume response data}
\format{
\subsection{\code{resume}}{

A data frame with 7,332 rows and 7 columns:
\describe{
\item{crime}{1 if applicant has criminal record, 0 otherwise}
\item{drugcrime}{1 if applicant has committed drug crime, 0 otherwise}
\item{propertycrime}{1 if applicant has committed property crime, 0 otherwise}
\item{ged}{1 if applicant has GED, 0 otherwise}
\item{empgap}{1 if applicant has a gap in employment, 0 otherwise}
\item{black}{1 if applicant is black, 0 otherwise}
\item{response}{1 if applicant received positive response, 0 otherwise}
}
}
}
\source{
\doi{10.7910/DVN/VPHMNT}
}
\usage{
resume
}
\description{
Data on responses to hypothetical resumes that were created for an experimental
study, based upon "Ban the Box, Criminal Records, and Racial Discrimination:
A Field Experiment" by Amanda Agan and Sonja Starr, 2018, Quarterly Journal
of Economics, 133: 191-235. This dataset considers only the subsample from
before the ban-the-box initiative.
}
\keyword{datasets}
