% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sp500}
\alias{sp500}
\title{Monthly returns data for S&P 500 companies}
\format{
\subsection{\code{sp500}}{

A data frame with 364 rows and 268 columns:
\describe{
\item{Date}{Date, as a string, indicating the endpoint of the month}
\item{IDX}{Monthly return for the S&P 500 index}
\item{AAPL, ABMD, ..., ZION}{Monthly company returns, where variable name is the company stock ticker symbol}
}
}
}
\source{
\url{https://finance.yahoo.com}
}
\usage{
sp500
}
\description{
Data on monthly returns for S&P 500 companies between Jan 1991 and Apr 2021
}
\keyword{datasets}
