% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{strikes}
\alias{strikes}
\title{Strike duration data}
\format{
\subsection{\code{strikes}}{

A data frame with 566 rows and 1 column:
\describe{
\item{duration}{Strike duration (in weeks)}
}
}
}
\source{
\url{https://cameron.econ.ucdavis.edu/mmabook/mmadata.html}
}
\usage{
strikes
}
\description{
Data on the length of worker contract strikes within U.S. manufacturing for
the period 1968-1976, based upon "The Duration of Contract strikes in
U.S. Manufacturing" by John Kennan, 1985, Journal of Econometrics, 28: 5-28.
}
\keyword{datasets}
