% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{website}
\alias{website}
\title{Website visitor arrival data}
\format{
\subsection{\code{website}}{

A data frame with 748 rows and 2 columns:
\describe{
\item{arrival}{Arrival time during the hour (in minutes)}
\item{time_since_last}{Time since last visitor (in minutes)}
}
}
}
\usage{
website
}
\description{
Data on the arrival time of website visitors during a specific hour for
a hypothetical website.
}
\keyword{datasets}
