% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{widgets}
\alias{widgets}
\title{Hypothetical data for widgets.com website}
\format{
\subsection{\code{widgets}}{

A data frame with 3,000 rows and 4 columns:
\describe{
\item{emailA}{1 if customer receives e-mail A, 0 otherwise}
\item{emailB}{1 if customer receives e-mail B, 0 otherwise}
\item{purchase}{1 if customer makes a purchase, 0 otherwise}
\item{amount}{Total purchase (in dollars)}
}
}
}
\usage{
widgets
}
\description{
Data on purchases for an e-mail experiment run by widgets.com
}
\keyword{datasets}
