% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cigdata}
\alias{cigdata}
\title{State-level cigarette price and tax data}
\format{
\subsection{\code{cigdata}}{

A data frame with 51 rows and 9 columns:
\describe{
\item{state}{State abbreviation}
\item{statename}{State name}
\item{price_pack}{Average price per pack (in dollars)}
\item{pack_sales_per_capita}{Annual sales, packs per capita}
\item{totaltax_pct}{Total tax (federal plus state) per pack, as percent of pack price}
\item{totaltax_pack}{Total tax (federal plus state) per pack (in dollars)}
\item{cig_tax_revnue}{Total annual tax revenue (in dollars)}
\item{producer}{1 if tobacco production > 20m pounds, 0 otherwise}
}
}
}
\source{
\url{https://healthdata.gov/dataset/The-Tax-Burden-on-Tobacco-1970-2019/etts-u9ii}
}
\usage{
cigdata
}
\description{
Data on cigarette prices and taxes in 2019 for the 50 U.S. states plus the
District of Columbia
}
\keyword{datasets}
